(wip) - keepup - add border to avatar to indicate urgency to keepup
The idea is to create an indicator that informs the user at a glance how soon or urgent it is to keep up with the contact.

See the 3rd resource below to create a gradient. I think if we can create a mixture of green and red and show more red as time goes on (untill fully red) - it visually conveys this message.

Resources:
https://medium.com/@mohammedijas/circle-image-avatar-with-border-in-flutter-513cdf82df43
https://vocal.media/futurism/mastering-circle-avatar-image-in-flutter
https://stackoverflow.com/questions/71135183/how-do-you-give-a-circleavatar-multiple-border-colors-in-flutter
https://pub.dev/packages/rainbow_color