keepup - UI tweaks background and colors
Overall we want to improve the design and layout of the app. This will be an initial effort that will likely require further polishing again at a later time.

- Let’s use a very light grey background.
- Text should be dark grey. This will be the primary text color. We will use other text colors, like a lighter grey and maybe something red or orange for alert text.

The issue is if we simply change the primaryColor to light grey then the app will change many places besides the background. See my screenshot below.

In the next tickets - we will apply styling changes target areas.

https://www.loom.com/share/8bae4975d39340188266596d1b751035

A little tweaking in the code I did - see the light grey background....