keepup - Master Categories and Groups
Currently we support the creation of groups and contacts. 3 groups are auto-created on launch and then any additional groups reside with the initial 3. Really, we want to consider the initial three groups to be the master categories for all groups. This is more in line with the logic of the intended use-cases. For example, I create a close relatives group under family, or I create a students of college x under network. The Master-Categories are:

 Family

Friends

Network (subject to name change) – previously ‘Business Associates’

Implement the following:

Updating the data structures and models to associate groups to a master category. Each group belongs to a single category

When creating a group – it must be assigned to a category

Add a segment control on views header to filter by all or individual categories – will be explained further in each view. The purpose is to limit the display of too many contacts.

 Consider an allowance for additional categories (maybe even turn on off hidden/inactive ones)