keepup - Tapping a contact - action sheet display tweaks
Description :
When tapping a contact and the action sheet displaying. Allow for when tapping on the avatar to then display the ability to add a photo.
Also add a photo icon (small) to the right of the avatar when tapped launches the add avatar control modal.

Reference Video : https://www.loom.com/share/d8061d6585bd445da2f8c110299f8904