keepup - today and upcoming view tweaks
Description

Today list should show all contacts that have 1 day until expiration AND expired (we might add a filter to see today versus expired).

Upcoming should contain one list total. It should NOT contain any contacts that are expiring today. It should be more of a simple list of all. The query and sorting logic are as follows. Display all contacts in the order of expiration ascending (1,2,3days etc). Excluding expired. Since expired will be on the today view.

https://www.loom.com/share/eb93dc71fd764ec5b764abadbc1b377b