keepup - a series of tweaks
comment out title on today view. (the please keeup up the…)

group title on contact list item should be a point smaller

keepup button on contact-list-item should be an icon button.

icon: recommend

color should be a lighter grey if possible

change the icon in the action sheet of contact-options for KeepUp to recommend

update the icons in the action sheet to outlined. I think we’ll change all but let’s begin with this for now.

upcoming view should only contain one list. Comment out the this week and this month and the ‘No Contacts’ messages. The idea is that the badge number delineates the expiration connotation

Filter out on the add contacts-to-groups that begin with a number (screenshot below)