keepup - some more UI tweaks
We are trying to refine the look and feel of the app. To make it more engaging

Revert the avatar display to showing the badge with a moon crescent.

Do not display the days (comment out) in the badge. Just a '!' if expired and the moon phase.

Instead of the keep up button on the lists (comment out) - add the text {x} days left as the trailing element. Eventually if > than a month is should say {x} months left

And the avatar in the action sheet should be larger

Also - on the action sheet move last KeepUp to the top item